+++ lib/Braintree/WebhookNotification.php
@@ -36,6 +36,7 @@ class WebhookNotification extends Base
     const PAYMENT_METHOD_CUSTOMER_DATA_UPDATED = 'payment_method_customer_data_updated';
     const PAYMENT_METHOD_REVOKED_BY_CUSTOMER = 'payment_method_revoked_by_customer';
     const RECIPIENT_UPDATED_GRANTED_PAYMENT_METHOD = 'recipient_updated_granted_payment_method';
+    const REFUND_FAILED = 'refund_failed';
     const SUBSCRIPTION_BILLING_SKIPPED = 'subscription_billing_skipped';
     const SUBSCRIPTION_CANCELED = 'subscription_canceled';
     const SUBSCRIPTION_CHARGED_SUCCESSFULLY = 'subscription_charged_successfully';