+++ tests/unit/PaymentMethodTest.php
@@ -52,7 +52,7 @@ class PaymentMethodTest extends Setup
                             'firstName', 'lastName', 'company', 'countryName',
                             'countryCodeAlpha2', 'countryCodeAlpha3', 'countryCodeNumeric',
                             'extendedAddress', 'locality', 'postalCode', 'region',
-                            'streetAddress', 'phoneNumber'],
+                            'streetAddress', 'phoneNumber', ['internationalPhone' => ['countryCode', 'nationalNumber']]],
                     ],
                 ]],
             ]],